+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/DisintegrateSpell.java
@@ -60,11 +60,11 @@ public class DisintegrateSpell extends Spell
 				LivingEntity li = (LivingEntity)targetEntity;
 				if (li instanceof Player)
 				{
-					li.damage(playerSpells.getPowerMultiplier() * playerDamage, player);
+					li.damage(playerSpells.getDamageMultiplier() * playerDamage, player);
 				}
 				else
 				{
-					li.damage(playerSpells.getPowerMultiplier() * entityDamage, player);
+					li.damage(playerSpells.getDamageMultiplier() * entityDamage, player);
 				}
 				castMessage("ZOT!");
 				return SpellResult.SUCCESS;